Moved SampleTrack and it's managers to common/models.py
